Rename NI_Switch to Switch, fix imports
